Exporter: Fix camera FOV when output is vertical
Blender uses the 'sensor_fit' parameter to determine how to fit the
sensor inside of the output image dimension. By default, it is set to
'AUTO' which swaps x and y FOV axis depending on which side of the image
is largest.

Previously, we did not take that parameter into account, resulting in
invalid results for tall images.