Exporter: Fix camera FOV when output is vertical (#27)
This PR fixes #25.

Blender uses the `sensor_fit` parameter to determine how to fit the
sensor inside of the output image dimension. By default, it is set to
`AUTO` which swaps x and y FOV axis depending on which side of the image
is largest.

Previously, we did not take that parameter into account, resulting in
invalid results for tall images.

More information on how this parameter is used by Blender can be found [here](https://docs.blender.org/manual/en/latest/render/cameras.html#camera).