Exporter: Fix specular/IOR export for Principled BSDFs (#33)
Previously, the Mitsuba `specular` parameter of Principled BSDFs was obtained from Blender's `specular` parameter. However, this is incorrect if the material has a transmission component, in which case the `IOR` parameter must be taken into account.

Moreover, Mitsuba does not handle problematic values well (`specular` = 0 and `eta` = 1). We therefore add an epsilon to avoid these singularities.