Installation: Define a fixed version of Mitsuba dependencies (#44)
When installing dependencies from pip, the latest package is always
installed. This can create bugs that cannot be catched by CI as they are
not controlled by the environment.

In order to mitigate this, we define a supported version that is
enforced in order for the add-on to register. Note that custom Mitsuba
versions are not affected by this and one can always use any custom
built versions of Mitsuba with the add-on.