Geometry export: use unique names for repeated materials
When a user assigns the same material to multiple material slots, the
previous logic would use the same name for both and therefore
overwrite the exported mesh file with the mesh part that appeared
last in the materials slot list.
This ensures that unique names are used for the output meshes.